requirements.txt
pandas
gunicorn
matplotlib==3.4.3
json5==0.9.6
shap
pickle-mixin==1.0.2
plotly==5.1.0
Pillow==8.3.1
urllib3==1.26.4
jsonschema==3.2.0
streamlit==0.89.0
scikit-learn
seaborn


